abstract out formatMilliseconds and format1d0
