+++ src/coreUtils/coreUtils.js
@@ -1,19 +1,10 @@
-import { padTime } from './padTime';
 import { splitTime } from './splitTime';
+import { format1d0 } from './format1d0';
 
 export function coreUtils() {
   const lodash = window._;
   const Flow = window.Flow;
   const moment = window.moment;
-  const formatMilliseconds = s => {
-    const _ref = splitTime(s);
-    const hrs = _ref[0];
-    const mins = _ref[1];
-    const secs = _ref[2];
-    const ms = _ref[3];
-    return `${padTime(hrs)}:${padTime(mins)}:${padTime(secs)}.${ms}`;
-  };
-  const format1d0 = n => Math.round(n * 10) / 10;
   const formatElapsedTime = s => {
     const _ref = splitTime(s);
     const hrs = _ref[0];
@@ -40,7 +31,6 @@ export function coreUtils() {
     return code;
   };
   Flow.Util = {
-    formatMilliseconds,
     formatElapsedTime,
     formatClockTime,
     multilineTextToHTML,
