+++ src/coreUtils/format1d0.js
@@ -1,0 +1,3 @@
+export function format1d0(n) {
+  return Math.round(n * 10) / 10;
+}
